requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backoff==2.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.2
click==8.1.6
coloredlogs==15.0.1
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.2
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.13
fastapi==0.99.1
ffmpy==0.3.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
gradio==3.38.0
gradio_client==0.2.10
h11==0.14.0
html2text==2020.1.16
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
importlib-resources==6.0.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.240
langsmith==0.0.14
linkify-it-py==2.0.2
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
parsel==1.8.1
Pillow==10.0.0
posthog==3.0.1
Protego==0.2.1
protobuf==4.23.4
pulsar-client==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
PyDispatcher==2.0.7
pydub==0.25.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
rpds-py==0.9.2
Scrapy==2.9.0
semantic-version==2.10.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tldextract==3.4.4
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.65.0
Twisted==22.10.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
w3lib==2.1.1
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zope.interface==6.0
